Added audio context wrapper
